feat: add auth hook
