+++ apps/frontend/src/app/(main)/map/page.tsx
@@ -3,21 +3,12 @@
 // biome-ignore lint: shut up biome
 import Map from "react-map-gl/maplibre";
 import "maplibre-gl/dist/maplibre-gl.css";
-import { auth } from "@/auth/client";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
 import LoadingPage from "@/components/pages/LoadingPage";
 import ErrorPage from "@/components/pages/ErrorPage";
+import UseAuth from "@/hooks/use-auth";
 
 export default function MapPage() {
-  const { data: session, isPending, error } = auth.useSession();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!isPending && !session) {
-      router.push("/sign-in");
-    }
-  }, [session, isPending, router]);
+  const { session, isPending, error } = UseAuth();
 
   if (isPending || !session) return <LoadingPage />;
   if (error) return <ErrorPage />;
