+++ apps/frontend/src/app/(main)/memories/page.tsx
@@ -1,18 +1,9 @@
-import { auth } from "@/auth/client";
 import ErrorPage from "@/components/pages/ErrorPage";
 import LoadingPage from "@/components/pages/LoadingPage";
-import { useRouter } from "next/router";
-import { useEffect } from "react";
+import UseAuth from "@/hooks/use-auth";
 
 export default function Memories() {
-  const { data: session, isPending, error } = auth.useSession();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!isPending && !session) {
-      router.push("/sign-in");
-    }
-  }, [session, isPending, router]);
+  const { session, isPending, error } = UseAuth();
 
   if (isPending || !session) return <LoadingPage />;
   if (error) return <ErrorPage />;
